+++ functions/js2jsx.js
@@ -1,4 +1,4 @@
-const { transform } = require("../src/functions/transformObj2Jsx.js");
+const { transform } = require("../src/functions/js2jsx.js");
 
 const headers = {
   "Access-Control-Allow-Origin": "*",
